chore: simple table row context menu builder
