+++ shared/src/containers/SimpleTable/SimpleTable.tsx
@@ -11,8 +11,8 @@ import {
   FilterFn,
   SortingFn,
   sortingFns,
-  RowData,
   Table,
+  RowData,
   OnChangeFn,
   RowSelectionState,
   functionalUpdate,
@@ -22,31 +22,21 @@ import { useVirtualizer } from '@tanstack/react-virtual'
 import clsx from 'clsx'
 import useRowKeydown, { RowKeyboardEvent } from './hooks/useRowKeydown'
 
-import { RankingInfo, rankItem, compareItems, rankings } from '@tanstack/match-sorter-utils'
+import { rankItem, compareItems, rankings } from '@tanstack/match-sorter-utils'
 import { useSimpleTableContext } from './context/SimpleTableContext'
 import { SimpleTableCellTemplate, SimpleTableCellTemplateProps } from './SimpleTableRowTemplate'
 import { EmptyPlaceholder } from '@shared/components'
 import { RowPinningState } from '@tanstack/react-table'
+import {
+  SimpleTableProps,
+  SimpleTableRow,
+  SimpleTableRowContextMenuBuilder,
+} from './SimpleTable.types'
+import { useCreateContextMenu } from '../ContextMenu'
 
-declare module '@tanstack/react-table' {
-  //add fuzzy filter to the filterFns
-  interface FilterFns {
-    fuzzy: FilterFn<unknown>
-  }
-  interface FilterMeta {
-    itemRank: RankingInfo
-  }
-
-  interface TableMeta<TData extends RowData> {
-    isExpandable?: boolean
-    isLoading?: boolean
-    children?: (
-      props: SimpleTableCellTemplateProps,
-      row: Row<TData>,
-      table: Table<SimpleTableRow>,
-    ) => JSX.Element
-    [key: string]: any
-  }
+const toMenuItems = (items: ReturnType<SimpleTableRowContextMenuBuilder>) => {
+  if (!items) return []
+  return Array.isArray(items) ? items.filter(Boolean) : [items]
 }
 
 // Define a custom fuzzy filter function that will apply ranking info to rows (using match-sorter utils)
@@ -94,60 +84,6 @@ const fuzzySort: SortingFn<any> = (rowA, rowB, columnId) => {
 
   // Provide an alphanumeric fallback for when the item ranks are equal
   return dir === 0 ? sortingFns.alphanumeric(rowA, rowB, columnId) : dir
-}
-
-export type RowItemData = {
-  id: string
-  name?: string | null
-  label?: string | null
-  subType?: string | null
-  [key: string]: any
-}
-
-export type SimpleTableRow = {
-  id: string
-  parentId?: string
-  name: string
-  label: string
-  parents?: string[]
-  icon?: string | null
-  iconColor?: string
-  iconFilled?: boolean
-  img?: string | null
-  imgShape?: 'square' | 'circle'
-  startContent?: JSX.Element
-  endContent?: JSX.Element
-  subRows: SimpleTableRow[]
-  data: RowItemData
-  isDisabled?: boolean
-  disabledMessage?: string
-  inactive?: boolean
-}
-
-export interface SimpleTableProps extends Omit<React.HTMLAttributes<HTMLDivElement>, 'children'> {
-  data: SimpleTableRow[]
-  isLoading: boolean
-  error?: string
-  isExpandable?: boolean // show expand/collapse icons
-  isMultiSelect?: boolean // enable multi-select with shift+click and ctrl/cmd+click
-  enableClickToDeselect?: boolean // allow deselecting a single selected row by clicking it again & clicking outside clears selection
-  enableNonFolderIndent?: boolean // indent non-folder rows to align with folder rows
-  forceUpdateTable?: any
-  globalFilter?: string
-  meta?: Record<string, any>
-  rowHeight?: number // height of each row, used for virtual scrolling
-  imgRatio?: number
-  onScrollBottom?: () => void // callback fired when scrolled to the bottom of the table
-  fitContent?: boolean
-  children?: (
-    props: SimpleTableCellTemplateProps,
-    row: Row<SimpleTableRow>,
-    table: Table<SimpleTableRow>,
-  ) => JSX.Element
-  pt?: {
-    cell?: Partial<SimpleTableCellTemplateProps>
-    row?: Partial<React.HTMLAttributes<HTMLTableRowElement>>
-  }
 }
 
 // Helper function to get row range for shift-selection
@@ -198,6 +134,7 @@ const SimpleTable: FC<SimpleTableProps> = ({
   rowHeight,
   imgRatio,
   onScrollBottom,
+  rowContextMenuBuilders = [],
   children,
   pt,
   fitContent,
@@ -214,6 +151,7 @@ const SimpleTable: FC<SimpleTableProps> = ({
   } = useSimpleTableContext()
   const lastSelectedIdRef = useRef<string | null>(null)
   const tableRef = useRef<Table<SimpleTableRow> | null>(null)
+  const [showRowContextMenu] = useCreateContextMenu()
 
   // stable data reference
   const tableData = useMemo(() => {
@@ -500,6 +438,7 @@ const SimpleTable: FC<SimpleTableProps> = ({
     getExpandedRowModel: getExpandedRowModel(),
     filterFromLeafRows: true,
     // debugTable: true,
+    // @ts-expect-error: updateData is not required
     meta: {
       isExpandable: !!isExpandable,
       isLoading: isLoading,
@@ -515,6 +454,44 @@ const SimpleTable: FC<SimpleTableProps> = ({
   }, [table])
 
   const { rows } = table.getRowModel()
+
+  const handleRowContextMenu = useCallback(
+    (rowId: string, rowIndex: number, e: ReactMouseEvent<HTMLTableRowElement>) => {
+      if (!tableRef.current) return
+      const row = rows[rowIndex]
+      if (!row || row.original === undefined) return
+
+      e.preventDefault()
+      e.stopPropagation()
+
+      const selectedIds = Object.keys(rowSelection)
+
+      const nextSelectedRows = selectedIds.includes(rowId) ? selectedIds : [rowId]
+
+      if (!selectedIds.includes(rowId)) {
+        // convert array to RowSelectionState object
+        const nextSelection: RowSelectionState = { [rowId]: true }
+        onRowSelectionChange?.(nextSelection)
+      }
+
+      const menuItems = rowContextMenuBuilders.flatMap((builder) =>
+        toMenuItems(
+          builder(e as any, {
+            rowId,
+            rowIndex,
+            row,
+            selectedRows: nextSelectedRows,
+            isSelected: nextSelectedRows.includes(rowId),
+          }),
+        ),
+      )
+
+      if (menuItems.length > 0) {
+        showRowContextMenu(e as any, menuItems)
+      }
+    },
+    [rowContextMenuBuilders, rows, rowSelection, onRowSelectionChange, showRowContextMenu],
+  )
 
   //The virtualizer needs to know the scrollable container element
   const tableContainerRef = useRef<HTMLDivElement>(null)
@@ -591,6 +568,7 @@ const SimpleTable: FC<SimpleTableProps> = ({
                   ref={measureElementRef} //measure dynamic row height
                   key={row.id}
                   id={row.id}
+                  onContextMenu={(e) => handleRowContextMenu(row.id, virtualRow.index, e)}
                   {...pt?.row}
                   style={{
                     transform: `translateY(${virtualRow.start}px)`, //this should always be a `style` as it changes on scroll
