+++ shared/src/containers/SimpleTable/index.ts
@@ -1,5 +1,6 @@
 import SimpleTable from './SimpleTable'
 export * from './SimpleTable'
+export * from './SimpleTable.types'
 export * from './context/SimpleTableContext'
 export * from './SimpleTable.styled'
 export * from './SimpleTableRowTemplate'
